refactor: Split out error code
