+++ src/lib.rs
@@ -1,5 +1,7 @@
 extern crate serde;
 extern crate serde_json;
 
+mod error;
+pub use error::*;
 mod cargo;
 pub use cargo::*;
